Implemented the cloth shader.
This ment adding two new shaders in cudaFile.cu, only_shadows_closest_hit_radiance, which is used for the floor plane,
and cloth_closest_hit_radiance, which is the actual cloth shader.
I also had to change mesh_intersect in triangle_mesh.cu to compute uv derivatives, since they are needed for the cloth shader.
Furthermore, there was some changes to optixProject.cpp in order to send parameters to the shader, and also in order to load
several obj files at once.